+++ guillotina/async_util.py
@@ -1,10 +1,6 @@
-from datetime import datetime
 from dateutil.tz import tzutc
-from guillotina import logger
-from guillotina import task_vars
-from guillotina.browser import View
+from guillotina import logger, task_vars
 from guillotina.db.transaction import Status
-from guillotina.exceptions import RequestNotFound
 from guillotina.exceptions import ServerClosingException
 from guillotina.exceptions import TransactionNotFound
 from guillotina.interfaces import IAsyncJobPool  # noqa
@@ -13,8 +9,9 @@ from guillotina.interfaces import IQueueUtility  # noqa
 from guillotina.transactions import get_tm
 from guillotina.transactions import get_transaction
 from guillotina.transactions import transaction
-from guillotina.utils import find_container
-from guillotina.utils import get_current_request
+from guillotina.utils import dump_task_vars
+from guillotina.utils import execute
+from guillotina.utils import load_task_vars
 
 import asyncio
 import typing
@@ -42,8 +39,11 @@ class QueueUtility(object):
         while True:
             got_obj = False
             try:
-                view, tm, txn = await self.queue.get()
+                func, tvars = await self.queue.get()
                 got_obj = True
+                load_task_vars(tvars)
+                txn = get_transaction()
+                tm = get_tm()
                 if txn is None or (
                     txn.status in (Status.ABORTED, Status.COMMITTED, Status.CONFLICT) and txn._db_conn is None
                 ):
@@ -52,20 +52,16 @@ class QueueUtility(object):
                     # still finishing current transaction, this connection
                     # will be cut off, so we need to wait until we no longer
                     # have an active transaction on the reqeust...
-                    await self.add((view, tm, txn))
+                    await self.add((func, tvars))
                     await asyncio.sleep(1)
                     continue
 
-                container = find_container(view.context)
-                if container is not None:
-                    task_vars.container.set(container)
-                with view.request, tm, txn:
-                    try:
-                        await view()
-                        await tm.commit(txn=txn)
-                    except Exception as e:
-                        logger.error("Exception on writing execution", exc_info=e)
-                        await tm.abort(txn=txn)
+                try:
+                    await func()
+                    await tm.commit(txn=txn)
+                except Exception as e:
+                    logger.error("Exception on writing execution", exc_info=e)
+                    await tm.abort(txn=txn)
             except (
                 RuntimeError,
                 SystemExit,
@@ -81,12 +77,8 @@ class QueueUtility(object):
                 self._exceptions = True
                 logger.error("Worker call failed", exc_info=e)
             finally:
-                task_vars.container.set(None)
                 if got_obj:
-                    try:
-                        view.request.execute_futures()
-                    except AttributeError:
-                        pass
+                    execute.execute_futures()
                     self.queue.task_done()
 
     @property
@@ -98,9 +90,7 @@ class QueueUtility(object):
         return self._total_queued
 
     async def add(self, view):
-        tm = get_tm()
-        txn = get_transaction()
-        await self.queue.put((view, tm, txn))
+        await self.queue.put((view, dump_task_vars()))
         self._total_queued += 1
         return self.queue.qsize()
 
@@ -108,24 +98,12 @@ class QueueUtility(object):
         pass
 
 
-class QueueObject(View):
-    def __init__(self, context, request):
-        # not sure if we need proxy object here...
-        # super(QueueObject, self).__init__(context, TransactionProxy(request))
-        super(QueueObject, self).__init__(context, request)
-        self.time = datetime.now(tz=_zone).timestamp()
-
-    def __lt__(self, view):
-        return self.time < view.time
-
-
 class Job:
     def __init__(
-        self, func: typing.Callable[[], typing.Coroutine], request=None, tm=None, args=None, kwargs=None
+        self, func: typing.Callable[[], typing.Coroutine], _task_vars=None, args=None, kwargs=None
     ) -> None:
         self._func = func
-        self._tm = tm
-        self._request = request
+        self._task_vars = _task_vars
         self._args = args
         self._kwargs = kwargs
 
@@ -134,8 +112,11 @@ class Job:
         return self._func
 
     async def run(self):
-        if self._request is not None:
-            async with self._tm, transaction(tm=self._tm), self._request:
+        if self._task_vars is not None:
+            load_task_vars(self._task_vars)
+        tm = task_vars.tm.get()
+        if tm is not None:
+            async with transaction(tm=tm):
                 await self._func(*self._args or [], **self._kwargs or {})
         else:
             # if no request, we do it without transaction
@@ -170,31 +151,22 @@ class AsyncJobPool:
     async def finalize(self):
         await self.join()
 
-    def add_job(self, func: typing.Callable[[], typing.Coroutine], request=None, args=None, kwargs=None):
+    def add_job(self, func: typing.Callable[[], typing.Coroutine], args=None, kwargs=None):
         if self._closing:
             raise ServerClosingException("Can not schedule job")
-        if request is None:
-            try:
-                request = get_current_request()
-            except RequestNotFound:
-                pass
-        job = Job(func, request=request, tm=get_tm(), args=args, kwargs=kwargs)
+        job = Job(func, _task_vars=dump_task_vars(), args=args, kwargs=kwargs)
         self._pending.insert(0, job)
         self._schedule()
         return job
 
-    def _add_job_after_commit(self, status, func, request=None, args=None, kwargs=None):
-        self.add_job(func, request=request, args=args, kwargs=kwargs)
+    def _add_job_after_commit(self, status, func, args=None, kwargs=None):
+        self.add_job(func, args=args, kwargs=kwargs)
 
-    def add_job_after_commit(
-        self, func: typing.Callable[[], typing.Coroutine], request=None, args=None, kwargs=None
-    ):
+    def add_job_after_commit(self, func: typing.Callable[[], typing.Coroutine], args=None, kwargs=None):
         txn = get_transaction()
         if txn is not None:
             txn.add_after_commit_hook(
-                self._add_job_after_commit,
-                args=[func],
-                kws={"request": request, "args": args, "kwargs": kwargs},
+                self._add_job_after_commit, args=[func], kws={"args": args, "kwargs": kwargs}
             )
         else:
             raise TransactionNotFound("Could not find transaction to run job with")